feat: implement support for weather sensors; related to #72
BREAKING CHANGE: weather sensors are untested due to lack of device
