--- .flakeheaven_cache/975e7a37850e323f5cb88e2c904ec06e.json
@@ -1,1 +1,0 @@
-{"results": [["pycodestyle", "W503", 282, 16, "line break before binary operator", "                + position_keyframe.duration\n"], ["pycodestyle", "W503", 283, 16, "line break before binary operator", "                + position_keyframe.delay\n"], ["pycodestyle", "W503", 306, 16, "line break before binary operator", "                or (position_end_time is None and current_time > angle_end_time)\n"], ["pycodestyle", "W503", 307, 16, "line break before binary operator", "                or (current_time > position_end_time and current_time > angle_end_time)\n"]], "digest": "dd31060492c965047fa90dcd5a336b37"}
