--- .flakeheaven_cache/e04f20f45df6269780bfca68615a220e.json
@@ -1,1 +1,0 @@
-{"results": [["pycodestyle", "W503", 156, 12, "line break before binary operator", "            | SUPPORT_CLOSE\n"], ["pycodestyle", "W503", 157, 12, "line break before binary operator", "            | SUPPORT_STOP\n"], ["pycodestyle", "W503", 158, 12, "line break before binary operator", "            | SUPPORT_SET_POSITION\n"], ["pycodestyle", "W503", 159, 12, "line break before binary operator", "            | SUPPORT_SET_TILT_POSITION\n"]], "digest": "c77c271510873caa3a064cd768028129"}
