+++ .github/workflows/verify.yml
@@ -20,7 +20,7 @@ jobs:
       max-parallel: 4
       matrix:
         python-version: ["3.11"]
-        poetry-version: [1.5.1]
+        poetry-version: [1.7.0]
 
     steps:
       - name: Checkout repository
@@ -64,7 +64,7 @@ jobs:
       max-parallel: 4
       matrix:
         python-version: ["3.11"]
-        poetry-version: [1.5.1]
+        poetry-version: [1.7.0]
 
     steps:
       - name: Checkout repository
