+++ custom_components/hella_onyx/enum/__init__.py
@@ -1,0 +1,1 @@
+"""The ONYX.CENTER Home Assistant enums."""
