+++ custom_components/hella_onyx/manifest.json
@@ -5,7 +5,7 @@
   "documentation": "https://github.com/muhlba91/onyx-homeassistant-integration",
   "issue_tracker": "https://github.com/muhlba91/onyx-homeassistant-integration/issues",
   "requirements": [
-    "onyx-client==7.1.1"
+    "onyx-client==7.2.1"
   ],
   "ssdp": [],
   "zeroconf": [],
