+++ custom_components/hella_onyx/sensors/__init__.py
@@ -1,0 +1,1 @@
+"""The ONYX.CENTER Home Assistant sensors."""
