+++ hacs.json
@@ -1,5 +1,5 @@
 {
   "name": "Hella ONYX.CENTER",
-  "homeassistant": "2023.8.0",
+  "homeassistant": "2023.11.0",
   "render_readme": true
 }
