+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.11"
-onyx-client = "^7.1.1"
+onyx-client = "^7.2.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.3.1"
@@ -39,7 +39,7 @@ black = "^23.3.0"
 pre-commit = "^3.3.2"
 pytest-asyncio = "^0.21.0"
 aioresponses = "^0.7.4"
-homeassistant = "^2023.8.0"
+homeassistant = "^2023.11.0"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
