+++ tests/test_sensor.py
@@ -1,4 +1,4 @@
-"""Test for the LED-Pi Sensor Entities."""
+"""Test for the ONYX Sensors."""
 
 from unittest.mock import MagicMock, patch
 
@@ -6,8 +6,8 @@ import pytest
 from homeassistant.config_entries import ConfigEntry
 from onyx_client.data.device_mode import DeviceMode
 from onyx_client.device.light import Light
+from onyx_client.device.weather import Weather
 from onyx_client.device.shutter import Shutter
-from onyx_client.enum.action import Action
 from onyx_client.enum.device_type import DeviceType
 
 from custom_components.hella_onyx import (
@@ -16,10 +16,7 @@ from custom_components.hella_onyx import (
     ONYX_COORDINATOR,
     ONYX_TIMEZONE,
 )
-from custom_components.hella_onyx.sensor import (
-    OnyxSensorDeviceType,
-    async_setup_entry,
-)
+from custom_components.hella_onyx.sensor import async_setup_entry
 
 
 @patch("homeassistant.core.HomeAssistant")
@@ -42,6 +39,13 @@ async def test_async_setup_entry(mock_hass):
             DeviceMode(DeviceType.BASIC_LIGHT),
             list(),
         ),
+        "weather": Weather(
+            "weather",
+            "name",
+            DeviceType.WEATHER,
+            DeviceMode(DeviceType.WEATHER),
+            list(),
+        ),
     }
     async_add_entries = AsyncAddEntries()
     mock_hass.data = {
@@ -56,8 +60,22 @@ async def test_async_setup_entry(mock_hass):
 
     await async_setup_entry(mock_hass, config_entry, async_add_entries.call)
     assert async_add_entries.called_async_add_entities
-    assert len(async_add_entries.data) == 1
+    assert len(async_add_entries.data) == 8
     assert async_add_entries.data[0]._uuid == "shutter"
+    assert async_add_entries.data[1]._uuid == "weather"
+    assert async_add_entries.data[1].unique_id == "weather/DeviceType"
+    assert async_add_entries.data[2]._uuid == "weather"
+    assert async_add_entries.data[2].unique_id == "weather/Humidity"
+    assert async_add_entries.data[3]._uuid == "weather"
+    assert async_add_entries.data[3].unique_id == "weather/Temperature"
+    assert async_add_entries.data[4]._uuid == "weather"
+    assert async_add_entries.data[4].unique_id == "weather/AirPressure"
+    assert async_add_entries.data[5]._uuid == "weather"
+    assert async_add_entries.data[5].unique_id == "weather/WindPeak"
+    assert async_add_entries.data[6]._uuid == "weather"
+    assert async_add_entries.data[6].unique_id == "weather/SunBrightnessPeak"
+    assert async_add_entries.data[7]._uuid == "weather"
+    assert async_add_entries.data[7].unique_id == "weather/SunBrightnessSink"
 
 
 @patch("homeassistant.core.HomeAssistant")
@@ -88,46 +106,6 @@ async def test_async_setup_entry_filter_all(mock_hass):
     await async_setup_entry(mock_hass, config_entry, async_add_entries.call)
     assert async_add_entries.called_async_add_entities
     assert len(async_add_entries.data) == 0
-
-
-class TestOnyxSensorDeviceType:
-    @pytest.fixture
-    def api(self):
-        yield MagicMock()
-
-    @pytest.fixture
-    def coordinator(self):
-        yield MagicMock()
-
-    @pytest.fixture
-    def device(self):
-        yield Shutter(
-            "id",
-            "name",
-            DeviceType.RAFFSTORE_90,
-            DeviceMode(DeviceType.RAFFSTORE_90),
-            list(Action),
-        )
-
-    @pytest.fixture
-    def entity(self, api, coordinator):
-        yield OnyxSensorDeviceType(
-            api, "UTC", coordinator, "name", DeviceType.RAFFSTORE_90, "uuid"
-        )
-
-    def test_icon(self, entity):
-        assert entity.icon == "mdi:cellphone-link"
-
-    def test_name(self, entity):
-        assert entity.name == "name Device Type"
-
-    def test_unique_id(self, entity):
-        assert entity.unique_id == "uuid/DeviceType"
-
-    def test_state(self, api, entity, device):
-        api.device.return_value = device
-        assert entity.state == DeviceType.RAFFSTORE_90.string()
-        assert api.device.called
 
 
 class AsyncAddEntries:
